Animation frame format change, add lua class for struct arrays

+++ src/zeq/animated_model.hpp
@@ -28,7 +28,7 @@ private:
 
     struct Frame
     {
-        uint32_t    milliseconds;
+        float       milliseconds;
         Vec3        pos;
         Quaternion  rot;
         Vec3        scale;
